Add patch to fix usage of reallocarray
Without _GNU_SOURCE defined at build time, the implicit declaration of
the reallocarray function ends up breaking the pointers returned,
causing among other things, the build of doxygen to fail with a pretty
segmentation fault.

The flex.license file was removed to let autospec run its magic, since
without that it was removing the BSD part, which seems to be the right
one looking at the project.

Build_pattern is added to force autospec to use autotools, since it's
finding an .indent.pro file in the source and assuming this means it's a
qmake based project. That also adds qtbase-dev and qtbase-extras as
dependencies for the build, so those are banned to avoid overhead while
building.